+++ src/main/java/com/macuguita/lib/impl/client/MacuLibClient.java
@@ -23,7 +23,7 @@ import org.jetbrains.annotations.ApiStatus;
 import com.macuguita.lib.api.event.player.client.ClientPlayerJoinEvent;
 import com.macuguita.lib.api.event.player.client.ClientPlayerLeaveEvent;
 import com.macuguita.lib.api.network.PacketRegistry;
-import com.macuguita.lib.impl.persista.S2CDataUpdatedPacket;
+import com.macuguita.lib.impl.persista.ClientboundDataUpdatedPacket;
 import com.macuguita.lib.impl.platform.ClientAbstraction;
 import com.macuguita.lib.impl.supporters.CapeManager;
 
@@ -36,7 +36,7 @@ public class MacuLibClient implements ClientModInitializer {
 
 		CapeManager.fetchAvailableCapes();
 
-		PacketRegistry.registerClientboundPacketHandler(S2CDataUpdatedPacket.TYPE, S2CDataUpdatedPacket::handle);
+		PacketRegistry.registerClientboundPacketHandler(ClientboundDataUpdatedPacket.TYPE, ClientboundDataUpdatedPacket::handle);
 	}
 
 	private void registerEvents() {