+++ CHANGELOG.md
@@ -1,1 +1,8 @@
+3.0.1
+- Fix some rate limiting problems with persista api
+- Update to 26.1
+---
+3.0.0
+- Rewrite the mod to be multiloader
+- Add persista api
 - Updated LICENSE